# CUDA: Fix #6339, performance regression launching specialized kernels