@gmarkall 2020-10-09
The fails here appear to be conda HTTP errors.