@stuartarchibald 2020-10-12
Thanks for the patch. For me, mainline gives:
```
90.9 µs ± 1.08 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
181 µs ± 1.61 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
256 µs ± 175 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
332 µs ± 12 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
402 µs ± 8.22 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
this patch gives:
```
64 µs ± 373 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
96.3 µs ± 385 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
120 µs ± 1.08 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
143 µs ± 115 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
165 µs ± 393 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```
on a GeForce GTX 750 Ti.