@stuartarchibald 2020-10-12
> Note: the force push is substituting "µ" for "u" in the commit message in order to enable the commit to be tested on platforms that don't support non-ASCII characters.

Many thanks @gmarkall