@stuartarchibald 2020-10-12
Unfortunately `±` is now causing issues