@gmarkall 2020-10-13
> Unfortunately `±` is now causing issues

Pushed again with this swapped for `+/-`. I've checked it's all ASCII this time!