@stuartarchibald 2020-10-13
> > Unfortunately `±` is now causing issues
> 
> Pushed again with this swapped for `+/-`. I've checked it's all ASCII this time!

Thanks!

Buildfarm ID: `numba_smoketest_cuda_120` is for 3b6a17e