@stuartarchibald 2020-10-13
> > > Unfortunately `±` is now causing issues
> > 
> > 
> > Pushed again with this swapped for `+/-`. I've checked it's all ASCII this time!
> 
> Thanks!
> 
> Buildfarm ID: `numba_smoketest_cuda_120` is for [3b6a17e](https://github.com/numba/numba/commit/3b6a17e643f54b707be8545c808f60c6d1a5e4ec)

This passed.